requirements.txt
PyYAML>=6.0.1
Jinja2>=3.1.4
fastapi
uvicorn[standard]
